Allow 'others' key in fill dicts
