+++ torchvision/transforms/v2/_utils.py
@@ -1,9 +1,7 @@
 import collections.abc
-import functools
 import numbers
-from collections import defaultdict
 from contextlib import suppress
-from typing import Any, Callable, Dict, Literal, Optional, Sequence, Type, TypeVar, Union
+from typing import Any, Callable, Dict, Literal, Optional, Sequence, Type, Union
 
 import torch
 
@@ -31,28 +29,11 @@ def _setup_float_or_seq(arg: Union[float, Sequence[float]], name: str, req_size:
 
 def _check_fill_arg(fill: Union[_FillType, Dict[Type, _FillType]]) -> None:
     if isinstance(fill, dict):
-        for key, value in fill.items():
-            # Check key for type
+        for value in fill.values():
             _check_fill_arg(value)
-        if isinstance(fill, defaultdict) and callable(fill.default_factory):
-            default_value = fill.default_factory()
-            _check_fill_arg(default_value)
     else:
         if fill is not None and not isinstance(fill, (numbers.Number, tuple, list)):
             raise TypeError("Got inappropriate fill arg, only Numbers, tuples, lists and dicts are allowed.")
-
-
-T = TypeVar("T")
-
-
-def _default_arg(value: T) -> T:
-    return value
-
-
-def _get_defaultdict(default: T) -> Dict[Any, T]:
-    # This weird looking construct only exists, since `lambda`'s cannot be serialized by pickle.
-    # If it were possible, we could replace this with `defaultdict(lambda: default)`
-    return defaultdict(functools.partial(_default_arg, default))
 
 
 def _convert_fill_arg(fill: datapoints._FillType) -> datapoints._FillTypeJIT:
@@ -74,13 +55,18 @@ def _setup_fill_arg(fill: Union[_FillType, Dict[Type, _FillType]]) -> Dict[Type,
     if isinstance(fill, dict):
         for k, v in fill.items():
             fill[k] = _convert_fill_arg(v)
-        if isinstance(fill, defaultdict) and callable(fill.default_factory):
-            default_value = fill.default_factory()
-            sanitized_default = _convert_fill_arg(default_value)
-            fill.default_factory = functools.partial(_default_arg, sanitized_default)
         return fill  # type: ignore[return-value]
+    else:
+        return {"others": _convert_fill_arg(fill)}
 
-    return _get_defaultdict(_convert_fill_arg(fill))
+
+def _get_fill(fill_dict, inpt_type):
+    if inpt_type in fill_dict:
+        return fill_dict[inpt_type]
+    elif "others" in fill_dict:
+        return fill_dict["others"]
+    else:
+        RuntimeError("This should never happen, please open an issue on the torchvision repo if you hit this.")
 
 
 def _check_padding_arg(padding: Union[int, Sequence[int]]) -> None:
